+++ src/main/java/de/unihalle/informatik/MiToBo/core/datatypes/images/MTBImage.java
@@ -484,7 +484,7 @@ public abstract class MTBImage extends ALDData
   /**
    * Set the physical size of a voxel (stepsize) in x-dimension
    * 
-   * @param stepsize
+   * @param stepsize	Stepsize in x dimension.
    */
   public void setStepsizeX(double stepsize) {
     this.calibration.pixelWidth = stepsize;
@@ -494,7 +494,7 @@ public abstract class MTBImage extends ALDData
   /**
    * Set the physical size of a voxel (stepsize) in y-dimension
    * 
-   * @param stepsize
+   * @param stepsize	Stepsize in y dimension.
    */
   public void setStepsizeY(double stepsize) {
 	this.calibration.pixelHeight = stepsize;
@@ -504,7 +504,7 @@ public abstract class MTBImage extends ALDData
   /**
    * Set the physical size of a voxel (stepsize) in z-dimension
    * 
-   * @param stepsize
+   * @param stepsize	Stepsize in z dimension.
    */
   public void setStepsizeZ(double stepsize) {
 	this.calibration.pixelDepth = stepsize;
@@ -514,7 +514,7 @@ public abstract class MTBImage extends ALDData
   /**
    * Set the stepsize in t-dimension (timestep)
    * 
-   * @param stepsize
+   * @param stepsize	Stepsize in t dimension.
    */
   public void setStepsizeT(double stepsize) {
     this.calibration.frameInterval = stepsize;
@@ -1406,6 +1406,45 @@ public abstract class MTBImage extends ALDData
   }
   
   /**
+   * Draws a 2D line into the given slice of the image.
+   * <p>
+   * This function basically relies on the Bresenham algorithm for rendering
+   * line segments as implemented in method
+   * {@link MTBLineSegment2D#getPixelsAlongSegment()} of class
+   * {@link MTBLineSegment2D}.
+   * 
+   * @param xstart	x-coordinate of start point.
+   * @param ystart	y-coordinate of start point.
+   * @param xend		x-coordinate of end point.
+   * @param yend		y-coordinate of end point.
+   * @param z       z-coordinate of image slice. 
+   * @param t 			t-coordinate of image slice.
+   * @param c 			c-coordinate of image slice.
+   * @param value		Color or gray-scale value to use for drawing the segment.
+   */
+  public void drawLine2D(int xstart, int ystart, int xend, int yend, 
+  		int z, int t, int c, int value) {
+
+  	MTBLineSegment2D line = new MTBLineSegment2D(xstart, ystart, xend, yend);
+  	
+  	LinkedList<Point2D.Double> pixelList = line.getPixelsAlongSegment();
+
+  	int x, y;
+  	for (Point2D.Double p: pixelList) {
+  		x = (int)p.x;
+  		y = (int)p.y;
+  		
+  		// check for pixel not falling outside of image domain
+      if (   x >= 0 && x < this.getSizeX() 
+      		&& y >= 0 && y < this.getSizeY()
+      		&& z >= 0 && z < this.getSizeZ()
+      		&& t >= 0 && t < this.getSizeT()
+      		&& c >= 0 && c < this.getSizeC())
+        this.putValueInt(x, y, z, t, c, value);
+  	}  	
+  }
+
+  /**
    * Draws a point at given position into the x-y-plane.
    * <p>
    * MTBImages are 5D, but here t- and c-dimensions are ignored.
@@ -1485,6 +1524,47 @@ public abstract class MTBImage extends ALDData
   	}
   }
   
+  /**
+   * Draws a 2D circle at the given z position into the x-y-plane.
+   * <p>
+   * MTBImages are 5D, but here t- and c-dimensions are ignored.
+   * @param x				Center of circle in x.
+   * @param y				Center of circle in y.
+   * @param z 			Circle position in z.
+   * @param radius 	Radius of circle.
+   * @param color 	Color of circle
+   */
+  public void drawFilledCircle2D(int x, int y, int z, int radius, int color) {
+  	int icx, icy;
+  	int squRadius = radius*radius;
+  	for (int dy=0; dy<=radius; ++dy) {
+    	for (int dx=0; dx<=radius; ++dx) {
+    		if ((dx*dx) + (dy*dy) <= squRadius) {
+    			// bottom-right
+    			icx = x+dx;
+    			icy = y+dy;
+      		if (icx < this.m_sizeX && icy < this.m_sizeY)
+      			this.putValueInt(icx, icy, z, 0, 0, color);
+      		// top-left
+    			icx = x-dx;
+    			icy = y-dy;
+      		if (icx >= 0 && icy >= 0)
+      			this.putValueInt(icx, icy, z, 0, 0, color);
+      		// bottom-left
+    			icx = x-dx;
+    			icy = y+dy;
+      		if (icx >= 0 && icy < this.m_sizeY)
+      			this.putValueInt(icx, icy, z, 0, 0, color);
+      		// top-right
+    			icx = x+dx;
+    			icy = y-dy;
+      		if (icx < this.m_sizeX && icy >= 0)
+      			this.putValueInt(icx, icy, z, 0, 0, color);
+    		}
+    	}
+  	}
+  }
+
   /**
    * Get the slice label of the slice specified by the current slice index
    * @return	Label of current slice.