+++ src/main/java/de/unihalle/informatik/MiToBo/features/regions/Region2DSkeletonAnalyzer.java
@@ -1,7 +1,7 @@
 /*
  * This file is part of MiToBo, the Microscope Image Analysis Toolbox.
  *
- * Copyright (C) 2010 - 2014
+ * Copyright (C) 2010 - @YEAR@
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by